Moved gui test to package
