+++ cmd/uelp/main.go
@@ -1,65 +1,10 @@
 package main
 
 import (
-	"github.com/gotk3/gotk3/gtk"
+	"github.com/sams96/uelp/pkg/gui/gtk"
 	"os"
 )
 
-const WindowName = "window"
-const UIMain = "glade/main.glade"
-
 func main() {
-	gtk.Init(&os.Args)
-	bldr, err := getBuilder(UIMain)
-	if err != nil {
-		panic(err)
-	}
-
-	window, err := getWindow(bldr)
-	if err != nil {
-		panic(err)
-	}
-
-	window.SetTitle("uelp")
-	window.SetDefaultSize(640, 480)
-	_, err = window.Connect("destroy", func() {
-		gtk.MainQuit()
-	})
-	if err != nil {
-		panic(err)
-	}
-
-	window.ShowAll()
-
-	gtk.Main()
-}
-
-func getBuilder(filename string) (*gtk.Builder, error) {
-	b, err := gtk.BuilderNew()
-	if err != nil {
-		return nil, err
-	}
-
-	if filename != "" {
-		err = b.AddFromFile(filename)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	return b, nil
-}
-
-func getWindow(b *gtk.Builder) (*gtk.Window, error) {
-	obj, err := b.GetObject(WindowName)
-	if err != nil {
-		return nil, err
-	}
-
-	window, ok := obj.(*gtk.Window)
-	if !ok {
-		return nil, err
-	}
-
-	return window, nil
+	gtk.Run(&os.Args)
 }
